feat: move helper components to sub dir
Helper component are those which don't need own route and used only in
example components.
