+++ src/playground/layout/infinite-list/components/news-post-placeholder.component.ts
@@ -1,0 +1,15 @@
+import { Component, HostBinding } from '@angular/core';
+
+@Component({
+  selector: 'nb-news-post-placeholder',
+  template: `
+    <div class="title-placeholder"></div>
+    <div class="text-placeholder"></div>
+    <div class="link-placeholder"></div>
+  `,
+  styleUrls: [ '../news-post-placeholder.component.scss' ],
+})
+export class NbNewsPostPlaceholderComponent {
+  @HostBinding('attr.aria-label')
+  label = 'Loading';
+}
