+++ src/playground/layout/tabset/components/route-tabset-children.component.ts
@@ -1,0 +1,25 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'nb-route-tabset-showcase-child1',
+  template: `<p>List of <strong>users</strong>.</p>`,
+  styles: [`
+    :host p {
+      padding: 1.25rem;
+    }
+  `],
+})
+export class NbRouteTabsetShowcaseChild1Component {
+}
+
+@Component({
+  selector: 'nb-route-tabset-showcase-child2',
+  template: `<p>List of <strong>orders</strong>.</p>`,
+  styles: [`
+    :host p {
+      padding: 1.25rem;
+    }
+  `],
+})
+export class NbRouteTabsetShowcaseChild2Component {
+}
