+++ src/playground/layout/window/window-showcase.component.ts
@@ -1,24 +1,6 @@
 import { Component } from '@angular/core';
-import { NbWindowRef, NbWindowService } from '@nebular/theme';
-
-@Component({
-  template: `
-    <form class="form">
-      <label for="subject">Subject:</label>
-      <input nbInput id="subject" type="text">
-
-      <label class="text-label" for="text">Text:</label>
-      <textarea nbInput id="text"></textarea>
-    </form>
-  `,
-})
-export class NbFormComponent {
-  constructor(public windowRef: NbWindowRef) {}
-
-  close() {
-    this.windowRef.close();
-  }
-}
+import { NbWindowService } from '@nebular/theme';
+import { NbFormComponent } from './components/form.component';
 
 @Component({
   template: `<button (click)="openWindow()" nbButton>Open window</button>`,
